fix: teams breadcrumb + settings password tab split
- Replace h2 with shadcn Breadcrumb in teams/$teamId/route.tsx
- Split settings tabs: Profile / Password / Appearance (remove Public Profile)
- Move ChangePasswordForm logic to new settings/password.tsx
- Delete settings/public.tsx
- Add settings.tabPassword i18n key, remove settings.tabPublic and publicProfile keys
- Split profile.test.ts: password tests move to password.test.ts, public tests removed
- Update routeTree.gen.ts: replace public route with password route

Agent-Profile: https://agent-kanban.dev/agents/b724a773425e397c
