+++ src/i18n/locales/en.json
@@ -236,8 +236,8 @@
   "admin.storages.customHostPlaceholder": "Optional",
   "settings.title": "Settings",
   "settings.tabProfile": "Profile",
+  "settings.tabPassword": "Password",
   "settings.tabAppearance": "Appearance",
-  "settings.tabPublic": "Public Profile",
   "settings.profile.section": "Profile",
   "settings.profile.displayName": "Display Name",
   "settings.profile.username": "Username",
@@ -256,8 +256,6 @@
   "settings.appearance.themeLight": "Light",
   "settings.appearance.themeDark": "Dark",
   "settings.appearance.language": "Language",
-  "settings.publicProfile.section": "Public Profile",
-  "settings.publicProfile.hint": "Share files to make them appear on your public profile.",
   "preview.download": "Download",
   "preview.close": "Close",
   "preview.unsupported": "Preview is not available for this file type.",
