+++ src/i18n/locales/zh.json
@@ -236,8 +236,8 @@
   "admin.storages.customHostPlaceholder": "可选",
   "settings.title": "设置",
   "settings.tabProfile": "基本信息",
+  "settings.tabPassword": "密码",
   "settings.tabAppearance": "外观",
-  "settings.tabPublic": "公开主页",
   "settings.profile.section": "个人资料",
   "settings.profile.displayName": "显示名称",
   "settings.profile.username": "用户名",
@@ -256,8 +256,6 @@
   "settings.appearance.themeLight": "浅色",
   "settings.appearance.themeDark": "深色",
   "settings.appearance.language": "语言",
-  "settings.publicProfile.section": "公开主页",
-  "settings.publicProfile.hint": "分享文件后，它们将出现在你的公开主页上。",
   "preview.download": "下载",
   "preview.close": "关闭",
   "preview.unsupported": "此文件类型暂不支持预览。",
