+++ src/routes/_authenticated/settings/profile.tsx
@@ -19,19 +19,7 @@ const profileSchema = z.object({
   displayName: z.string().min(1).max(100),
 })
 
-const passwordSchema = z
-  .object({
-    currentPassword: z.string().min(1),
-    newPassword: z.string().min(8),
-    confirmPassword: z.string().min(1),
-  })
-  .refine((data) => data.newPassword === data.confirmPassword, {
-    message: 'passwords_mismatch',
-    path: ['confirmPassword'],
-  })
-
 type ProfileFormValues = z.infer<typeof profileSchema>
-type PasswordFormValues = z.infer<typeof passwordSchema>
 
 export function ProfileForm() {
   const { t } = useTranslation()
@@ -85,68 +73,14 @@ export function ProfileForm() {
   )
 }
 
-export function ChangePasswordForm() {
-  const { t } = useTranslation()
-
-  const form = useForm<PasswordFormValues>({
-    resolver: zodResolver(passwordSchema),
-    defaultValues: { currentPassword: '', newPassword: '', confirmPassword: '' },
-  })
-
-  const mutation = useMutation({
-    mutationFn: async (values: PasswordFormValues) => {
-      const { error } = await authClient.changePassword({
-        currentPassword: values.currentPassword,
-        newPassword: values.newPassword,
-      })
-      if (error) throw error
-    },
-    onSuccess: () => {
-      toast.success(t('settings.profile.passwordChanged'))
-      form.reset()
-    },
-    onError: (err) => toast.error(err.message ?? String(err)),
-  })
-
-  return (
-    <form onSubmit={form.handleSubmit((v) => mutation.mutate(v))} className="space-y-4">
-      <div className="space-y-1.5">
-        <Label htmlFor="currentPassword">{t('settings.profile.currentPassword')}</Label>
-        <Input id="currentPassword" type="password" {...form.register('currentPassword')} />
-      </div>
-
-      <div className="space-y-1.5">
-        <Label htmlFor="newPassword">{t('settings.profile.newPassword')}</Label>
-        <Input id="newPassword" type="password" {...form.register('newPassword')} />
-      </div>
-
-      <div className="space-y-1.5">
-        <Label htmlFor="confirmPassword">{t('settings.profile.confirmPassword')}</Label>
-        <Input id="confirmPassword" type="password" {...form.register('confirmPassword')} />
-        {form.formState.errors.confirmPassword && (
-          <p className="text-xs text-destructive">{t('settings.profile.passwordMismatch')}</p>
-        )}
-      </div>
-
-      <Button type="submit" disabled={mutation.isPending}>
-        {mutation.isPending ? t('common.loading') : t('settings.profile.changePassword')}
-      </Button>
-    </form>
-  )
-}
-
 function ProfilePage() {
   const { t } = useTranslation()
 
   return (
-    <div className="max-w-lg space-y-6">
+    <div className="max-w-lg">
       <div className="space-y-4 rounded-md border p-4">
         <h3 className="text-sm font-medium text-muted-foreground">{t('settings.profile.section')}</h3>
         <ProfileForm />
-      </div>
-      <div className="space-y-4 rounded-md border p-4">
-        <h3 className="text-sm font-medium text-muted-foreground">{t('settings.profile.changePassword')}</h3>
-        <ChangePasswordForm />
       </div>
     </div>
   )
